fix: create separate `toArray` util for `packages/nuxt/src/app`
